# contracts/ai.openapi.json
{
  "components": {
    "schemas": {
      "AssistantRequest": {
        "additionalProperties": false,
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "reasonCode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          }
        },
        "required": [
          "question",
          "locale",
          "service",
          "page"
        ],
        "title": "AssistantRequest",
        "type": "object"
      },
      "AssistantResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "blockedReason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation"
          },
          "fallbackUsed": {
            "title": "Fallbackused",
            "type": "boolean"
          },
          "sourceIds": {
            "items": {
              "type": "string"
            },
            "title": "Sourceids",
            "type": "array"
          },
          "uncertain": {
            "title": "Uncertain",
            "type": "boolean"
          }
        },
        "required": [
          "answer",
          "uncertain",
          "fallbackUsed"
        ],
        "title": "AssistantResponse",
        "type": "object"
      },
      "BlockedReason": {
        "enum": [
          "PII_DETECTED",
          "UNSUPPORTED_REQUEST",
          "PROVIDER_UNAVAILABLE",
          "SAFETY_CHECK_UNAVAILABLE"
        ],
        "title": "BlockedReason",
        "type": "string"
      },
      "Locale": {
        "enum": [
          "en",
          "hi"
        ],
        "title": "Locale",
        "type": "string"
      },
      "Page": {
        "enum": [
          "dashboard",
          "learner_application",
          "learner_test",
          "waiting_period",
          "driving_application",
          "appointment_search",
          "appointment_waitlist",
          "appointment_offer",
          "appointment_confirmed",
          "service_form"
        ],
        "title": "Page",
        "type": "string"
      },
      "ReasonCode": {
        "enum": [
          "ACTION_LOCKED",
          "WAITING_PERIOD_ACTIVE",
          "LEARNER_LICENCE_EXPIRING",
          "NO_MATCHING_SLOT",
          "WAITLIST_ACTIVE",
          "OFFER_EXPIRING",
          "SIMULATED_ACTION",
          "CHECKLIST_REQUIRED"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
      "Service": {
        "enum": [
          "learner_licence",
          "driving_licence",
          "licence_renewal",
          "duplicate_licence",
          "address_change",
          "mobile_update",
          "application_status",
          "fees_and_payment",
          "appointment_booking"
        ],
        "title": "Service",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Private, stateless explanation service for the DigiLicense prototype",
    "title": "DigiLicense AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "liveness_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Liveness Health Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "Not configured"
          }
        },
        "summary": "Readiness",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/assistant/messages": {
      "post": {
        "operationId": "assistant_message_v1_assistant_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid service credentials"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Service is not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Message",
        "tags": [
          "assistant"
        ]
      }
    }
  }
}