# services/ai/app/models/enums.py
"""Allowlisted values crossing or describing the AI trust boundary."""

from enum import StrEnum


class Locale(StrEnum):
    ENGLISH = "en"
    HINDI = "hi"


class Service(StrEnum):
    LEARNER_LICENCE = "learner_licence"
    DRIVING_LICENCE = "driving_licence"
    LICENCE_RENEWAL = "licence_renewal"
    DUPLICATE_LICENCE = "duplicate_licence"
    ADDRESS_CHANGE = "address_change"
    MOBILE_UPDATE = "mobile_update"
    APPLICATION_STATUS = "application_status"
    FEES_AND_PAYMENT = "fees_and_payment"
    APPOINTMENT_BOOKING = "appointment_booking"


class Page(StrEnum):
    DASHBOARD = "dashboard"
    LEARNER_APPLICATION = "learner_application"
    LEARNER_TEST = "learner_test"
    WAITING_PERIOD = "waiting_period"
    DRIVING_APPLICATION = "driving_application"
    APPOINTMENT_SEARCH = "appointment_search"
    APPOINTMENT_WAITLIST = "appointment_waitlist"
    APPOINTMENT_OFFER = "appointment_offer"
    APPOINTMENT_CONFIRMED = "appointment_confirmed"
    SERVICE_FORM = "service_form"


class ReasonCode(StrEnum):
    ACTION_LOCKED = "ACTION_LOCKED"
    WAITING_PERIOD_ACTIVE = "WAITING_PERIOD_ACTIVE"
    LEARNER_LICENCE_EXPIRING = "LEARNER_LICENCE_EXPIRING"
    NO_MATCHING_SLOT = "NO_MATCHING_SLOT"
    WAITLIST_ACTIVE = "WAITLIST_ACTIVE"
    OFFER_EXPIRING = "OFFER_EXPIRING"
    SIMULATED_ACTION = "SIMULATED_ACTION"
    CHECKLIST_REQUIRED = "CHECKLIST_REQUIRED"


class Intent(StrEnum):
    CURRENT_STEP = "CURRENT_STEP"
    LOCKED_ACTION = "LOCKED_ACTION"
    WAITING_PERIOD = "WAITING_PERIOD"
    LEARNER_LICENCE_EXPIRY = "LEARNER_LICENCE_EXPIRY"
    NO_APPOINTMENT_AVAILABLE = "NO_APPOINTMENT_AVAILABLE"
    WAITLIST_EXPLANATION = "WAITLIST_EXPLANATION"
    OFFER_EXPIRY = "OFFER_EXPIRY"
    MOCK_VS_REAL = "MOCK_VS_REAL"
    PREPARATION_CHECKLIST = "PREPARATION_CHECKLIST"
    UNSUPPORTED = "UNSUPPORTED"


class DLPAction(StrEnum):
    ALLOW = "ALLOW"
    BLOCK_PII = "BLOCK_PII"
    UNSUPPORTED = "UNSUPPORTED"
    FALLBACK = "FALLBACK"


class PIIEntityType(StrEnum):
    AADHAAR = "AADHAAR"
    PAN = "PAN"
    PASSPORT = "PASSPORT"
    VOTER_ID = "VOTER_ID"
    PHONE_NUMBER = "PHONE_NUMBER"
    OTP = "OTP"
    LICENCE_NUMBER = "LICENCE_NUMBER"
    APPLICATION_NUMBER = "APPLICATION_NUMBER"
    RECEIPT_NUMBER = "RECEIPT_NUMBER"
    VEHICLE_REGISTRATION = "VEHICLE_REGISTRATION"
    UPI_ID = "UPI_ID"
    IFSC = "IFSC"
    BANK_ACCOUNT = "BANK_ACCOUNT"
    PAYMENT_CARD = "PAYMENT_CARD"
    ADDRESS = "ADDRESS"
    PAYMENT_INFORMATION = "PAYMENT_INFORMATION"
    IDENTITY_INFORMATION = "IDENTITY_INFORMATION"


class BlockedReason(StrEnum):
    PII_DETECTED = "PII_DETECTED"
    UNSUPPORTED_REQUEST = "UNSUPPORTED_REQUEST"
    PROVIDER_UNAVAILABLE = "PROVIDER_UNAVAILABLE"
    SAFETY_CHECK_UNAVAILABLE = "SAFETY_CHECK_UNAVAILABLE"